Separated dummy DP9800 from real. Added logging and a method to capture errors
